requirements.txt
pandas==2.0.1
PyYAML==6.0.1
PyYAML==6.0.1
streamlit==1.31.1
streamlit_authenticator==0.3.1
streamlit_extras==0.4.0
streamlit_modal==0.1.2
yfinance==0.2.18
